backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "!"
authors = ["Ivan Perelygin <jperelygin@gmail.com>"]
packages = [{ include = "backend", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "*"
sqlalchemy = ">=2.0.46,<3.0.0"
alembic = ">=1.18.3,<2.0.0"
pydantic = ">=2.12.5,<3.0.0"
python-jose = ">=3.5.0,<4.0.0"
passlib = ">=1.7.4,<1.8.0"
bcrypt = ">=4.1.2,<4.2.0"
jinja2 = ">=3.1.6,<4.0.0"
uvicorn = "^0.40.0"
psycopg = {extras = ["binary"], version = "^3.3.2"}
mypy = "^1.19.1"
pydantic-settings = "^2.12.0"
email-validator = "^2.3.0"
requests = "^2.32.5"
httpx = "^0.28.1"
slowapi = "^0.1.9"
python-multipart = ">=0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"
pytest-asyncio = ">=1.3.0,<2.0.0"
sqlalchemy = {version = "^2.0.46", extras = ["mypy"]}
types-passlib = "^1.7.7.20260211"
types-python-jose = "^3.5.0.20250531"
types-jinja2 = "^2.11.9"

[tool.mypy]
python_version = "3.12"
files = "src"
strict = true
show_error_codes = true
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
plugins = [
    "sqlalchemy.ext.mypy.plugin"
]

[[tool.mypy.overrides]]
module = [
    "alembic.*",
    "pytest_asyncio.*",
    "slowapi.*",
    "limits.*"
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"